homework/car_dealer/requirements/dev.txt
-r base.txt

pytest-django==3.9.0
mock==4.0.2
factory-boy==2.12.0
